Finished sign up with email-password (without transition to other page).
