--- FlatFleet/CreateAnAccountPage.xaml.cs
@@ -1,11 +1,0 @@
-namespace FlatFleet
-{
-	public partial class CreateAnAccountPage : ContentPage
-	{
-		public CreateAnAccountPage()
-		{
-			InitializeComponent();
-			BindingContext = new CreateAnAccountViewModel();
-        }
-	}
-}
