+++ FlatFleet/Features/SignUp/SignUpView.xaml
@@ -127,7 +127,12 @@
                 </Label.GestureRecognizers>
             </Label>
         </Grid>
-        <Button Text="Continue" Background="{StaticResource AddButtonBlue}" HorizontalOptions="FillAndExpand" Margin ="20" VerticalOptions="Center" />
+        <Button 
+            Text="Continue" 
+            Background="{StaticResource AddButtonBlue}" 
+            HorizontalOptions="FillAndExpand" Margin ="20" 
+            VerticalOptions="Center" 
+            Command="{Binding SignUpCommand}"/>
         <Grid HorizontalOptions="Center" ColumnSpacing="5">
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="Auto"></ColumnDefinition>
