+++ FlatFleet/ViewModels/PrivacyPolicyViewModel.cs
@@ -12,7 +12,7 @@ namespace FlatFleet.ViewModels
         }
         private async void OnSignIn()
         {
-            await NavigationService.NavigateTo(typeof(CreateAnAccountPage));
+            await NavigationService.NavigateTo(typeof(SingInPage));
         }
     }
 }
