+++ FlatFleet/ViewModels/SignInViewModel.cs
@@ -35,7 +35,8 @@ namespace FlatFleet.ViewModels
 
         private async void OnCreateAccount()
         {
-            await NavigationService.NavigateTo(typeof(CreateAnAccountPage));
+            // await Shell.Current.GoToAsync("//SignUp");
+            await NavigationService.NavigateTo(typeof(SignUpPage));
         }
         private async void OnForgotPassword()
         {
